+++ packages/frontend/src/features/sync/components/syncModalFormContext.ts
@@ -1,0 +1,35 @@
+import { createFormContext } from '@mantine/form';
+
+export interface SyncModalFormValues {
+    name: string;
+    cron: string;
+    timezone: string | undefined;
+    options: {
+        gdriveId: string;
+        gdriveName: string;
+        gdriveOrganizationName: string;
+        url: string;
+        tabName: string;
+    };
+    saveInNewTab: boolean;
+}
+
+export const [
+    SyncModalFormProvider,
+    useSyncModalFormContext,
+    useSyncModalForm,
+] = createFormContext<SyncModalFormValues>();
+
+export const DEFAULT_VALUES: SyncModalFormValues = {
+    name: '',
+    cron: '0 9 * * *',
+    timezone: undefined,
+    options: {
+        gdriveId: '',
+        gdriveName: '',
+        gdriveOrganizationName: '',
+        url: '',
+        tabName: '',
+    },
+    saveInNewTab: false,
+};